frontend/requirements.txt
Flask
python-dotenv
Werkzeug